Feat/feed pagination (#74)
* feat: 게시글 서버 페이지네이션 구현

* feat: 프론트엔드 게시글 목록 조회 연동 및 좋아요 색 표시 구현

+++ backend/src/main/java/com/team05/petmeeting/domain/feed/dto/FeedListRes.java
@@ -2,7 +2,6 @@ package com.team05.petmeeting.domain.feed.dto;
 
 import com.team05.petmeeting.domain.feed.entity.Feed;
 import com.team05.petmeeting.domain.feed.enums.FeedCategory;
-
 import java.time.LocalDateTime;
 
 public record FeedListRes(
@@ -19,7 +18,7 @@ public record FeedListRes(
         LocalDateTime createdAt,
         LocalDateTime updatedAt
 ) {
-    public FeedListRes(Feed feed, int likeCount, boolean isLiked) {
+    public FeedListRes(Feed feed, Long likeCount, boolean isLiked) {
         this(
                 feed.getId(),
                 feed.getUser() != null ? feed.getUser().getId() : null,
@@ -28,7 +27,7 @@ public record FeedListRes(
                 feed.getTitle(),
                 feed.getContent(),
                 feed.getImageUrl(),
-                likeCount,
+                likeCount.intValue(),
                 isLiked,
                 feed.getComments().size(),
                 feed.getCreatedAt(),
